optimise proxy_pass memory size a little, also boxing requestcontext since it is large
